format new files , update home.js
