+++ coffee-snob-app/components/cafe.js
@@ -1,27 +1,8 @@
 import React from 'react';
 import CafeCard from './CafeCard';
 import { FlatList, StyleSheet, View } from 'react-native';
+import cafe from '../data/CafeData'
 
-export const cafe = [
-    {
-        image: require('../assets/botanic.jpeg'),
-        name: 'Botanic Brews',
-        description: 'Tranquil oasis for plant lovers and coffee enthusiasts alike.',
-
-    },
-    {
-        image: require('../assets/rustic.avif'),
-        name: 'Barnwood Bistro',
-        description: 'Charming rustic cafe nestled in the heart of the countryside.',
-        
-    },
-    {
-        image: require('../assets/white.jpeg'),
-        name: 'Marble Cafe',
-        description: 'Sophisticated sanctuary known for its elegant white marble aesthetic.',
-        
-    }
-]
 
 const CafeList = () => {
     return (
