+++ coffee-snob-app/data/CafeData.js
@@ -1,0 +1,30 @@
+export const cafe = [
+    {
+        image: require('../assets/botanic.jpeg'),
+        name: 'Botanic Brews',
+        description: 'Tranquil oasis for plant lovers and coffee enthusiasts alike.',
+        rating: 4, 
+        openingHours: '8:00 AM - 6:00 PM', 
+        location: '123 Plant Street, City', 
+
+    },
+    {
+        image: require('../assets/rustic.avif'),
+        name: 'Barnwood Bistro',
+        description: 'Charming rustic cafe nestled in the heart of the countryside.',
+        rating: 4, 
+        openingHours: '8:00 AM - 6:00 PM', 
+        location: '123 Plant Street, City', 
+    },
+    {
+        image: require('../assets/white.jpeg'),
+        name: 'Marble Cafe',
+        description: 'Sophisticated sanctuary known for its elegant white marble aesthetic.',
+        rating: 4, 
+        openingHours: '8:00 AM - 6:00 PM', 
+        location: '123 Plant Street, City', 
+    }
+]
+
+
+export default cafe;
